requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.28.0
peft>=0.10.0

# Unsloth for efficient training
unsloth[colab-new]>=2024.5
unsloth_zoo>=2024.5

# HuggingFace ecosystem
huggingface_hub>=0.22.0
safetensors>=0.4.0

# Experiment tracking
wandb>=0.16.0

# Quantization
bitsandbytes>=0.43.0
auto-gptq>=0.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
psutil>=5.9.0

# Optional: for advanced features
deepspeed>=0.14.0
flash-attn>=2.5.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
